+++ attributes/server.rb
@@ -1,3 +1,4 @@
+# Encoding: utf-8
 default['logstash']['server']['version'] = '1.3.2'
 default['logstash']['server']['home'] = "#{node['logstash']['basedir']}/server"
 default['logstash']['server']['log_file'] = '/var/log/logstash/server.log'
@@ -9,7 +10,7 @@ default['logstash']['server']['config_dir'] = 'etc/conf.d'
 default['logstash']['server']['config_file'] = 'logstash.conf'
 default['logstash']['server']['config_templates'] = []
 default['logstash']['server']['config_templates_cookbook'] = 'logstash'
-default['logstash']['server']['config_templates_variables'] = { }
+default['logstash']['server']['config_templates_variables'] = {}
 default['logstash']['server']['base_config'] = 'server.conf.erb' # set blank if don't want data driven config
 default['logstash']['server']['base_config_cookbook'] = 'logstash'
 default['logstash']['server']['xms'] = '1024M'
@@ -35,4 +36,4 @@ default['logstash']['server']['web']['enable']  = false
 default['logstash']['server']['web']['address'] = '0.0.0.0'
 default['logstash']['server']['web']['port']    = '9292'
 
-default['logstash']['server']['logrotate']['options'] = [ 'missingok', 'notifempty', 'compress', 'copytruncate' ]
+default['logstash']['server']['logrotate']['options'] = %w{ missingok notifempty compress copytruncate }