+++ recipes/agent.rb
@@ -1,13 +1,14 @@
+# Encoding: utf-8
 #
 # Cookbook Name:: logstash
 # Recipe:: agent
 #
 #
-include_recipe "java"
-include_recipe "logstash::default"
+include_recipe 'logstash::default'
+include_recipe 'yum::default'
 
 if node['logstash']['agent']['init_method'] == 'runit'
-  include_recipe "runit"
+  include_recipe 'runit'
   service_resource = 'runit_service[logstash_agent]'
 else
   service_resource = 'service[logstash_agent]'
@@ -20,29 +21,8 @@ else
 end
 
 if node['logstash']['install_zeromq']
-  case
-  when platform_family?("rhel")
-    include_recipe "yumrepo::zeromq"
-  when platform_family?("debian")
-    apt_repository "zeromq-ppa" do
-      uri "http://ppa.launchpad.net/chris-lea/zeromq/ubuntu"
-      distribution node['lsb']['codename']
-      components ["main"]
-      keyserver "keyserver.ubuntu.com"
-      key "C7917B12"
-      action :add
-    end
-    apt_repository "libpgm-ppa" do
-      uri "http://ppa.launchpad.net/chris-lea/libpgm/ubuntu"
-      distribution  node['lsb']['codename']
-      components ["main"]
-      keyserver "keyserver.ubuntu.com"
-      key "C7917B12"
-      action :add
-      notifies :run, "execute[apt-get update]", :immediately
-    end
-  end
-  node['logstash']['zeromq_packages'].each {|p| package p }
+  include_recipe 'logstash::zero_mq_repo'
+  node['logstash']['zeromq_packages'].each { |p| package p }
 end
 
 # check if running chef-solo.  If not, detect the logstash server/ip by role.  If I can't do that, fall back to using ['logstash']['agent']['server_ipaddress']
@@ -50,7 +30,7 @@ if Chef::Config[:solo]
   logstash_server_ip = node['logstash']['agent']['server_ipaddress']
 else
   logstash_server_results = search(:node, "roles:#{node['logstash']['agent']['server_role']}")
-  unless logstash_server_results.empty?
+  if !logstash_server_results.empty?
     logstash_server_ip = logstash_server_results[0]['ipaddress']
   else
     logstash_server_ip = node['logstash']['agent']['server_ipaddress']
@@ -59,7 +39,7 @@ end
 
 directory node['logstash']['agent']['home'] do
   action :create
-  mode "0755"
+  mode '0755'
   owner node['logstash']['user']
   group node['logstash']['group']
 end
@@ -67,26 +47,26 @@ end
 %w{bin etc lib tmp log}.each do |ldir|
   directory "#{node['logstash']['agent']['home']}/#{ldir}" do
     action :create
-    mode "0755"
+    mode '0755'
     owner node['logstash']['user']
     group node['logstash']['group']
   end
 
-  link "/var/lib/logstash/#{ldir}" do
+  link "#{node['logstash']['homedir']}/#{ldir}" do
     to "#{node['logstash']['agent']['home']}/#{ldir}"
   end
 end
 
 directory "#{node['logstash']['agent']['home']}/etc/conf.d" do
   action :create
-  mode "0755"
+  mode '0755'
   owner node['logstash']['user']
   group node['logstash']['group']
 end
 
 directory patterns_dir do
   action :create
-  mode "0755"
+  mode '0755'
   owner node['logstash']['user']
   group node['logstash']['group']
 end
@@ -97,17 +77,17 @@ node['logstash']['patterns'].each do |file, hash|
     source 'patterns.erb'
     owner node['logstash']['user']
     group node['logstash']['group']
-    variables( :patterns => hash )
+    variables(:patterns => hash)
     mode '0644'
     notifies :restart, service_resource
   end
 end
 
-if node['logstash']['agent']['install_method'] == "jar"
+if node['logstash']['agent']['install_method'] == 'jar'
   remote_file "#{node['logstash']['agent']['home']}/lib/logstash-#{node['logstash']['agent']['version']}.jar" do
-    owner "root"
-    group "root"
-    mode "0755"
+    owner 'root'
+    group 'root'
+    mode '0755'
     source node['logstash']['agent']['source_url']
     checksum  node['logstash']['agent']['checksum']
     action :create_if_missing
@@ -118,7 +98,7 @@ if node['logstash']['agent']['install_method'] == "jar"
     notifies :restart, service_resource
   end
 else
-  include_recipe "logstash::source"
+  include_recipe 'logstash::source'
 
   logstash_version = node['logstash']['source']['sha'] || "v#{node['logstash']['server']['version']}"
   link "#{node['logstash']['agent']['home']}/lib/logstash.jar" do
@@ -127,28 +107,27 @@ else
   end
 end
 
-if node['logstash']['agent']['config_file']
-  template "#{node['logstash']['agent']['home']}/#{node['logstash']['agent']['config_dir']}/#{node['logstash']['agent']['config_file']}" do
-    source node['logstash']['agent']['base_config']
-    cookbook node['logstash']['agent']['base_config_cookbook']
-    owner node['logstash']['user']
-    group node['logstash']['group']
-    mode "0644"
-    variables(
-            :logstash_server_ip => logstash_server_ip,
-            :patterns_dir => patterns_dir)
-    notifies :restart, service_resource
-  end
+template "#{node['logstash']['agent']['home']}/#{node['logstash']['agent']['config_dir']}/#{node['logstash']['agent']['config_file']}" do
+  source node['logstash']['agent']['base_config']
+  cookbook node['logstash']['agent']['base_config_cookbook']
+  owner node['logstash']['user']
+  group node['logstash']['group']
+  mode '0644'
+  variables(
+          :logstash_server_ip => logstash_server_ip,
+          :patterns_dir => patterns_dir)
+  notifies :restart, service_resource
+  only_if { node['logstash']['agent']['config_file'] }
 end
 
-unless node['logstash']['agent']['config_templates'].empty? or node['logstash']['agent']['config_templates'].nil?
-  node['logstash']['agent']['config_templates'].each do |config_template| 
+unless node['logstash']['agent']['config_templates'].empty? || node['logstash']['agent']['config_templates'].nil?
+  node['logstash']['agent']['config_templates'].each do |config_template|
     template "#{node['logstash']['agent']['home']}/#{node['logstash']['agent']['config_dir']}/#{config_template}.conf" do
       source "#{config_template}.conf.erb"
       cookbook node['logstash']['agent']['config_templates_cookbook']
       owner node['logstash']['user']
       group node['logstash']['group']
-      mode "0644"
+      mode '0644'
       variables node['logstash']['agent']['config_templates_variables'][config_template]
       notifies :restart, service_resource
       action :create
@@ -159,36 +138,56 @@ end
 log_dir = ::File.dirname node['logstash']['agent']['log_file']
 directory log_dir do
   action :create
-  mode "0755"
+  mode '0755'
   owner node['logstash']['user']
   group node['logstash']['group']
   recursive true
 end
 
 if node['logstash']['agent']['init_method'] == 'runit'
-  runit_service "logstash_agent"
+  runit_service 'logstash_agent'
 elsif node['logstash']['agent']['init_method'] == 'native'
-  if platform_family? "debian"
-    if node["platform_version"] >= "12.04"
-      template "/etc/init/logstash_agent.conf" do
-        mode "0644"
-        source "logstash_agent.conf.erb"
+  if platform_family? 'debian'
+    if node['platform_version'] >= '12.04'
+      template '/etc/init/logstash_agent.conf' do
+        mode '0644'
+        source 'logstash_agent.conf.erb'
         notifies :restart, service_resource
       end
 
-      service "logstash_agent" do
+      service 'logstash_agent' do
         provider Chef::Provider::Service::Upstart
-        action [ :enable, :start ]
+        action [:enable, :start]
       end
     else
       Chef::Log.fatal("Please set node['logstash']['agent']['init_method'] to 'runit' for #{node['platform_version']}")
     end
-  elsif platform_family? "rhel", "fedora"
-    template "/etc/init.d/logstash_agent" do
-      source "init.logstash_server.erb"
-      owner "root"
-      group "root"
-      mode "0774"
+  elsif platform_family? 'fedora' && node['platform_version'] >= '15'
+    execute 'reload-systemd' do
+      command 'systemctl --system daemon-reload'
+      action :nothing
+    end
+
+    template '/etc/systemd/system/logstash_agent.service' do
+      source 'logstash_agent.service.erb'
+      owner 'root'
+      group 'root'
+      mode  '0755'
+      notifies :run, 'execute[reload-systemd]', :immediately
+      notifies :restart, 'service[logstash_agent]', :delayed
+    end
+
+    service 'logstash_agent' do
+      service_name 'logstash_agent.service'
+      provider Chef::Provider::Service::Systemd
+      action [:enable, :start]
+    end
+  elsif platform_family? 'rhel', 'fedora'
+    template '/etc/init.d/logstash_agent' do
+      source 'init.logstash_server.erb'
+      owner 'root'
+      group 'root'
+      mode '0774'
       variables(
         :config_file => node['logstash']['agent']['config_dir'],
         :home => node['logstash']['agent']['home'],
@@ -199,7 +198,7 @@ elsif node['logstash']['agent']['init_method'] == 'native'
       )
     end
 
-    service "logstash_agent" do
+    service 'logstash_agent' do
       supports :restart => true, :reload => true, :status => true
       action :enable
     end
@@ -208,7 +207,7 @@ else
   Chef::Log.fatal("Unsupported init method: #{node['logstash']['server']['init_method']}")
 end
 
-logrotate_app "logstash" do
+logrotate_app 'logstash' do
   path "#{log_dir}/*.log"
   if node['logstash']['logging']['useFileSize']
     size node['logstash']['logging']['maxSize']
@@ -218,7 +217,7 @@ logrotate_app "logstash" do
   rotate node['logstash']['logging']['maxBackup']
   options node['logstash']['agent']['logrotate']['options']
   create "664 #{node['logstash']['user']} #{node['logstash']['group']}"
-  notifies :restart, "service[rsyslog]"
+  notifies :restart, 'service[rsyslog]'
   if node['logstash']['agent']['logrotate']['stopstartprepost']
     prerotate <<-EOF
       service logstash_agent stop
@@ -230,4 +229,3 @@ logrotate_app "logstash" do
     EOF
   end
 end
-