+++ recipes/source.rb
@@ -1,10 +1,10 @@
-include_recipe "build-essential"
-include_recipe "java"
-include_recipe "ant"
-include_recipe "git"
-include_recipe "logstash::default"
+# Encoding: utf-8
+include_recipe 'build-essential'
+include_recipe 'ant'
+include_recipe 'git'
+include_recipe 'logstash::default'
 
-package "wget"
+package 'wget'
 
 logstash_version = node['logstash']['source']['sha'] || "v#{node['logstash']['server']['version']}"
 
@@ -12,7 +12,7 @@ directory "#{node['logstash']['basedir']}/source" do
   action :create
   owner node['logstash']['user']
   group node['logstash']['group']
-  mode "0755"
+  mode '0755'
 end
 
 git "#{node['logstash']['basedir']}/source" do
@@ -23,10 +23,12 @@ git "#{node['logstash']['basedir']}/source" do
   group node['logstash']['group']
 end
 
-execute "build-logstash" do
+execute 'build-logstash' do
   cwd "#{node['logstash']['basedir']}/source"
-  environment ({'JAVA_HOME' => node['logstash']['source']['java_home']})
-  user "root"
+  environment(
+    :JAVA_HOME => node['logstash']['source']['java_home']
+  )
+  user 'root'
   # This variant is useful for troubleshooting stupid environment problems
   # command "make clean && make VERSION=#{logstash_version} --debug > /tmp/make.log 2>&1"
   command "make clean && make VERSION=#{logstash_version} jar"