+++ test/unit/spec/beaver_spec.rb
@@ -1,0 +1,14 @@
+# Encoding: utf-8
+require_relative 'spec_helper'
+
+describe 'logstash::beaver' do
+  before { logstash_stubs }
+  describe 'ubuntu' do
+    let(:chef_run) { ChefSpec::Runner.new.converge(described_recipe) }
+
+    it 'writes some chefspec code' do
+      pending 'todo'
+    end
+
+  end
+end