+++ CHANGELOG.md
@@ -2,8 +2,25 @@
 
 This file is used to list changes made in each version of chef-logstash.
 
+## 0.7.7
+* Support for new beaver config [#239](https://github.com/lusis/chef-logstash/pull/239)
+* Support for multiline codec [#240](https://github.com/lusis/chef-logstash/pull/240)
+* Parameterize /var/lib/logstash [#242](https://github.com/lusis/chef-logstash/pull/242)
+* Fix parameter spacing option [#244](https://github.com/lusis/chef-logstash/pull/244)
+
+## 0.7.6:
+* introduced more testing
+  * Strainer: rubocop, knife test, foodcritic, chefspec
+  * lots of style fixes for rubocop
+  * skeleton spec files for each recipe
+  * testkitchen + server spec
+
+## 0.7.5:
+* added fedora systemd support
+* moved zeromq repos to own recipe
+
 ## 0.7.4:
-* bump logstash version to 1.3.1
+* bump logstash version to 1.3.2
 
 ## 0.7.3:
 * support for sudo with upstart for agent and server